--- src/pages/Layout/NavBar/NavBar.jsx
@@ -1,9 +1,0 @@
-function NavBar() {
-  return (
-    <div>
-      <h3>This is the navigation bar</h3>
-    </div>
-  );
-}
-
-export default NavBar;
