+++ src/views/login/components/OTPLogin.jsx
@@ -210,7 +210,7 @@ export default function OTPLogin() {
                 </Link>
               </Grid>
               <Grid item>
-                <Link href="#" variant="body2">
+                <Link href="/register" variant="body2">
                   {"Don't have an account? Sign Up"}
                 </Link>
               </Grid>
