+++ src/views/login/components/PasswordLogin.jsx
@@ -130,7 +130,7 @@ export default function PasswordLogin() {
                                 </Link>
                             </Grid>
                             <Grid item>
-                                <Link href="#" variant="body2">
+                                <Link href="/register" variant="body2">
                                     {"Don't have an account? Sign Up"}
                                 </Link>
                             </Grid>
